Rename srssl to rcst
