+++ rcst/proto/grsim/grSim_Robotstatus.proto
