+++ rcst/proto/referee/ssl_gc_common.proto
