+++ rcst/proto/referee/ssl_gc_game_event.proto
