+++ rcst/proto/referee/ssl_gc_geometry.proto
