+++ rcst/proto/vision/ssl_vision_geometry.proto
