+++ rcst/world_observer.py
